requirements.txt
aiogram==2.13
aiohttp==3.7.4.post0
altgraph==0.17
aniso8601==9.0.1
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.1
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
Babel==2.9.1
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
cfgv==3.2.0
chardet==3.0.4
choicelib==0.1.4
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
DAWG-Python==0.7.2
decorator==4.4.2
discord.py==1.6.0
distlib==0.3.4
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
dnspython==2.1.0
docopt==0.6.2
email-validator==1.1.2
filelock==3.0.12
Flask==1.1.2
Flask-Login==0.5.0
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
future==0.18.2
futures==3.1.1
googletrans==3.0.0
goslate==1.5.1
greenlet==1.0.0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
humanize==3.4.1
hyperframe==5.2.0
identify==2.2.4
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jwt==1.2.0
llvmlite==0.35.0
loguru==0.5.3
lxml==4.6.3
MarkupSafe==1.1.1
moviepy==1.0.3
multidict==5.1.0
nodeenv==1.6.0
numba==0.52.0
numpy==1.19.3
opencv-python==4.5.3.56
packaging==20.9
pandas==1.3.4
paramiko==2.8.0
pefile==2019.4.18
Pillow==8.1.0
pluggy==0.13.1
pre-commit==2.12.1
prettytable==2.1.0
proglog==0.1.9
psycopg2==2.8.6
py==1.10.0
pycparser==2.20
pydantic==1.8.2
PyExecJS==1.5.1
pygame==2.0.0
pyglet==1.5.14
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.11
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
PyNaCl==1.4.0
pyparsing==2.4.7
PyQt5==5.15.4
pyqt5-plugins==5.15.1.2.0.1
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyqt5-tools==5.15.1.3
pyqtgraph==0.12.1
pytest==6.2.2
python-box==5.3.0
python-dateutil==2.8.2
python-dotenv==0.15.0
python-telegram-bot==13.4.1
pytube==10.6.0
pytz==2021.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
qrcode==6.1
qt5-applications==5.15.2.2.1
qt5-tools==5.15.1.1.0.1
requests==2.25.1
rfc3986==1.4.0
schedule==1.0.0
scipy==1.7.1
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.4.5
SQLAlchemy-serializer==1.3.4.4
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
tox==3.23.0
tqdm==4.55.2
translate==3.5.0
translators==4.7.16
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
vbml==1.1
virtualenv==20.4.4
vk-api==11.9.1
vkbottle @ https://github.com/timoniq/vkbottle/archive/master.zip
vkbottle-types==0.1.24
voluptuous==0.12.1
watchgod==0.7
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.2.0
win32-setctime==1.0.3
WTForms==2.3.3
xlrd==2.0.1
XlsxWriter==1.3.7
yandex-weather-api==0.2.4
yarl==1.6.3
youtube-dl==2021.3.14
